+++ packages/capacitor/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - support Nx 12
 - plugin rewritten with `@nrwl/devkit` for better maintainability and future proofing for future Nx versions
+- update Capacitor to 3.2.5
 
 ## BREAKING CHANGES